Update to support latest env var syntax
@devTristan convinced me that `Secretfile` should support arbitrary
environment variables, so I updated emk/credentials to support this,
using a more Unix-like syntax:

    $VAULT_ENV
    ${VAULT_ENV}

The latter is useful when the variable appears in the middle of
something looks like an identifier:

    secrets/${VAULT_ENV}_postgresql

There are some other subtle enhancements to the format, but this is the
only one that might break compatibility.

If possible, could you please merge this (with any appropriate tweaks)
before your gem gets publically announced?